upload?

+++ db/migrate/20150526060211_create_exposure_gls.rb
@@ -9,8 +9,8 @@ class CreateExposureGls < ActiveRecord::Migration
       t.string :code, default: ""
       t.integer :premium_basis, default: 0
       t.string :type, default: ""
-      t.integer :base_rate, default: 0
-      t.integer :ilf, default: 0
+      t.decimal :base_rate, default: 0
+      t.decimal :ilf, default: 0
       t.integer :premium, default: 0
 
       t.timestamps null: false
